requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic-settings==2.1.0
python-telegram-bot==20.5
requests==2.31.0
beautifulsoup4==4.12.3
google-cloud-translate==3.11.2